requirements.txt
python-telegram-bot~=12.5.1
sqlalchemy~=1.3.15
colorthief~=0.2.1
colormath~=3.0.0
schedule~=0.6.0
sqlalchemy-mixins~=1.2.1
matplotlib~=3.3.2
psycopg2-binary~=2.8.6
